-Optimized SignalR code for presence of the user -Message group for the users is now optimized and properly marks messages as "read"

+++ DatingApp.Api/SignalR/MessageHub.cs
@@ -21,16 +21,19 @@ public class MessageHub(IMessageRepository messageRepository, IUserRepository us
         }
         var groupName = GetGroupName(Context.User.GetUsername(), otherUser);
         await Groups.AddToGroupAsync(Context.ConnectionId, groupName);
-        await AddToGroup(groupName);
+        var group = await AddToGroup(groupName);
+
+        await Clients.Group(groupName).SendAsync("UpdatedGroup", group);
 
         var messages = await messageRepository.GetMessageThread(Context.User.GetUsername(), otherUser!);
 
-        await Clients.Group(groupName).SendAsync("ReceiveMessageThread", messages);
+        await Clients.Caller.SendAsync("ReceiveMessageThread", messages);
     }
 
     public override async Task OnDisconnectedAsync(Exception? exception)
     {
-        await RemoveFromMessageGroup();
+        var group = await RemoveFromMessageGroup();
+        await Clients.Group(group.Name).SendAsync("UpdatedGroup", group);
         await base.OnDisconnectedAsync(exception);
     }
 
@@ -62,7 +65,7 @@ public class MessageHub(IMessageRepository messageRepository, IUserRepository us
         var groupName = GetGroupName(sender.UserName, recipient.UserName);
         var group = await messageRepository.GetMessageGroup(groupName);
 
-        if(group != null && group.Connections.Any(x => x.Username == recipient.UserName))
+        if (group != null && group.Connections.Any(x => x.Username == recipient.UserName))
         {
             message.DateRead = DateTime.UtcNow;
         }
@@ -82,30 +85,40 @@ public class MessageHub(IMessageRepository messageRepository, IUserRepository us
         }
     }
 
-    private async Task<bool> AddToGroup(string groupName)
+    private async Task<Group> AddToGroup(string groupName)
     {
         var username = Context.User?.GetUsername() ?? throw new Exception("Cannot get username");
         var group = await messageRepository.GetMessageGroup(groupName);
-        var connection = new Connection {ConnectionId = Context.ConnectionId, Username = username};
+        var connection = new Connection { ConnectionId = Context.ConnectionId, Username = username };
 
-        if(group == null)
+        if (group == null)
         {
             group = new Group { Name = groupName };
             messageRepository.AddGroup(group);
         }
 
         group.Connections.Add(connection);
-        return await messageRepository.SaveAllAsync();
+        if (await messageRepository.SaveAllAsync())
+        {
+            return group;
+        }
+        throw new HubException("Failed to join group");
     }
 
-    private async Task RemoveFromMessageGroup()
+    private async Task<Group> RemoveFromMessageGroup()
     {
-        var connection = await messageRepository.GetConnection(Context.ConnectionId);
-        if(connection != null)
+        var group = await messageRepository.GetGroupForConnection(Context.ConnectionId);
+        var connection = group?.Connections.FirstOrDefault(x => x.ConnectionId == Context.ConnectionId);
+        if (connection != null && group != null)
         {
             messageRepository.RemoveConnection(connection);
-            await messageRepository.SaveAllAsync();
+            if(await messageRepository.SaveAllAsync())
+            {
+                return group;
+            };
         }
+
+        throw new Exception("Failed to remove from group");
     }
 
     private string GetGroupName(string caller, string? other)
